Three more from the same sweep: a crash, a dead trap and a narrowing cast
Found while checking whether anything in this area still wanted a second pass.
None of them are new; all three were sitting behind the ones already fixed.

**Crafting threw instead of failing.** ProduceItem indexed [0] straight off the
list AddItem returns. AddItem delegates to TryAddItem, which returns an empty
list when it refuses the add, and it does refuse: it will not merge into a stack
that has gone from the data layer. So the single path where crafting cannot
deliver raised an index out of range rather than reporting a failure, and the
ingredients had already been consumed by then. It reports the failure and logs
which recipe.

**A dead method that was the wrong answer waiting to be picked up.**
GetUnequipped(Guid itemId) matched on the item id alone, which is exactly the
fault that lost tags and enchantments when TryAddItem used it to choose a stack
to merge into. TryAddItem asks GetUnequipped(StackKey) now and nothing else ever
called it, so it is deleted rather than left in place as the convenient wrong
thing to reach for next time.

**A silent narrowing on a purchase.** BuyMarketItem returned int while every
amount around it is a long, so the count of items a player had just paid for was
truncated with an unchecked cast on the way out. Amounts that large are not
realistic for a single listing, which is the argument for leaving it and not a
good one.

Suite unchanged at 46 passing, same two unrelated failures.

+++ src/RavenNest.BusinessLogic/Game/Managers/MarketplaceManager.cs
@@ -380,7 +380,16 @@ namespace RavenNest.BusinessLogic.Game
                 boughtTotalCost);
         }
 
-        private int BuyMarketItem(
+        /// <summary>
+        ///     How many were actually bought, which is zero when the purchase was refused.
+        /// </summary>
+        /// <remarks>
+        ///     Returned an int while every amount around it is a long, so the value was narrowed
+        ///     on the way out with an unchecked cast. Amounts that large are not realistic for one
+        ///     listing, but a silent wrap on the number of items a player just paid for is not a
+        ///     thing to leave in place on the grounds that it probably will not happen.
+        /// </remarks>
+        private long BuyMarketItem(
             SessionToken token,
             Guid itemId,
             Character character,
@@ -481,7 +490,7 @@ namespace RavenNest.BusinessLogic.Game
 
             AddGameEvent(token.SessionId, GameEventType.ItemBuy, model);
 
-            return (int)buyAmount;
+            return buyAmount;
         }
 
         private Character GetCharacter(SessionToken token, string userId, string platform)
